refactor: refactor bookings
