+++ app/dashboard/bookings/utils/dateUtils.ts
@@ -1,0 +1,15 @@
+export const formatDateTime = (dateString: string) => {
+  const date = new Date(dateString)
+  return {
+    date: date.toLocaleDateString('vi-VN', { day: '2-digit', month: '2-digit', year: 'numeric' }),
+    time: date.toLocaleTimeString('vi-VN', { hour: '2-digit', minute: '2-digit' }),
+    full: date.toLocaleString('vi-VN')
+  }
+}
+
+export const getDayOfWeek = (dateString: string) => {
+  const date = new Date(dateString)
+  const days = ['Chủ nhật', 'Thứ hai', 'Thứ ba', 'Thứ tư', 'Thứ năm', 'Thứ sáu', 'Thứ bảy']
+  return days[date.getDay()]
+}
+
